Replace 2D grid containers with 1D shelf containers
To simplify the hierarchy, I'm going to try following Tmux and other
tiling window managers (I believe!) by using a tree of 1D shelves in
place of the grid.

Adding a cell in the principal direction extends the current shelf,
while adding a cell in the opposing direction creates a child shelf.

So far this is feeling pretty intuitive, better than the hierarchy of
grids to me.